Try to fix "random death on self" bug
Looking carefully at what happened when this bug occurred, I discovered
that it did so when going at a close to vertical angle and turning in the
opposite direction just as the drawn x value increased. This should go for
a near horizontal angle when the drawn y value increased, as well.

What happened was that the Kurve would paint on a kuxel it had recently
drawn, but it would not understand that it is allowed to do so, and
therefore die. This turned out to happen because its this.secondLastDraw and
this.thirdLastDraw were never actually updated because of a reference error.