Fix major tickrate/draw issues
It would sometimes happen that a player would die just before actually
bumping into anything. The reason for this was that maxTicksBeforeDraw was
incorrectly calculated. I suspect that the reason why this would occur
pretty seldom (I'm taking a blind guess at a 4/64 change), and why the
error would never be more than 1 Kuxel, is that the max allowed framerate
(60 fps) was very close to the Kurve speed (64 Kx/s).

This commit rethinks the update/draw loop such that the max number of
ticks between draws is calculated as the tickrate over the speed, e.g.
600 Hz / 64 Kx/s = 9.375, rounded down to the closest integer, or 1 if <1.