Fix transform + flex bug
transform does not affect layout, so for example scale(2) in combination
with justify-content: flex-end would result in everything floating outside
of the viewport (to the right). This was fixed by setting
transform-origin: right, so that #wrapper is scaled "towards the left".
However, this is not desired when the field should be centered, so we also
had to set transform-origin: center in those cases.