Fix scaling pixelmapping issue
When in 2x mode, Chrome would display everything without any pixelmapping
at all. The reason for this is that the width of #wrapper was 721px -- an
odd number, more generally denoted by 2n+1. Scaling this by an integer
factor k gives a scaled width of (2n+1)k.

Because of transform-origin: center, this new width is always centered
relative to the original width 2n+1. This means that on each side (left
and right), there will be

((2n+1)k - (2n+1))/2 = (2nk + k - 2n - 1)/2 = nk + k/2 - n - 1/2 = n(k-1)
+ (k-1)/2

new pixels. If that is not an integer, we're in trouble! n(k-1) is always
an integer, so that's fine, but (k-1)/2 is only an integer when k is odd!
So we will never achieve pixelmapping for even scaling factors (e.g. k =
2).

This horrible bug was fixed by making #left 76px wide so that #wrapper is
instead 720px wide, which makes much more sense than 721px. No need to
overcomplicate anything!