Remove slow `let` compound statements in Game
V8 does not support optimization of compound assignments on variables
declared with `let` at the time of this commit (Chrome 49.0.2623). For
this reason, we will not use them in performance-critical functions, since
the entire function will be left unoptimized if it contains an unsupported
construct.

We won't use:

    let x = a;
    x += b;

We will instead use:

    let x = a;
    x = x + b;

https://jsperf.com/let-compound-assignment