Remove imperative live and gameOver properties
Instead of setting a `live` state imperatively, `isLive()` will check the
actual criteria for a Game being live (such as number of live players > 1
for a competitive game) whenever it is called. Unless that constitutes a
considerable performance hit, we prefer this approach because it is so
much easier to understand, verify and debug.

The same goes for `gameOver`.