Intercept everything, not just 'special' keys
Up until this commit, only explicitly designated special keys would be
intercepted and .preventDefault()ed. This does not cover multikey
shortcuts such as Alt + Left Arrow, which is unacceptable because the game
will be unloaded if for example Green and Yellow turn at the same time.

This commit removes the notion of 'special' keys and intercepts all key
events (except those outside the browser's control, such as Alt + Tab or
Win).