Add hole functionality
The players now create holes as they should, but their heads are not drawn
during the hole, which makes the game feel very laggy. This will be solved
in another commit in the near future.

When I implemented the hole functionality, I discovered that the hitbox
got completely fucked up during holes, so a player would die for
apparently no reason at all. The reason for this was that a Player's
`lastDraw` property would not be updated during a hole. That property is
now called `lastPosition` and is properly updated even if the player is
holy.