Add heads functionality; rewrite Renderer
The public API of Renderer and its internal translation are slightly
rewritten to fit into the concept of one main and one overlay canvas.

The canvases have gotten new IDs to better represent their roles.

Heads of Kurves are always drawn, even when the Kurve is holy. This is
absolutely essential for a playable experience.