Add mouse functionality
I'm not entirely satisfied with this solution, as it's kinda hacky:

MOUSE is like KEY; it holds keys that map to numbers representing mouse
buttons. But MOUSE.RIGHT is not 2; it is 1002. All numbers are offset by
1000 so as not to overlap the KEY key codes.

Therefore, when handling a mouse event, if one wants to compare the
clicked button to player configs, one must use `MOUSE.pack(event.button)`
rather than `event.button` as is.