Fix spawn holy bug
Sometimes a player would spawn and appear to create a ridiculously long
hole before starting to draw normally. I managed to pinpoint this to the
player keeping its `holy` state from the last round, i.e. if a player died
while holy, it would spawn holy in the next round.

The `holy` property is now properly set to `false` when calling `reset()`
on a `Player`. This is very imperative and not quite what I'd prefer, but
it works for now.