Fix multiple heads bug
If the hole size was large so that the lone head was clearly visible, one
could clearly see that there would often be multiple overlapping heads
next to each other. The reason for this was that the head of a player
would be drawn for every queued draw on said player, and there can be many
queued draws per refresh.

Now, the head is only drawn once per refresh (when all queued draws have
been processed, so always at the most recent position).