Improve implementation of hole size and interval
The earlier implementation would interpret a hole size of 8 as "8 Kuxels
from the center of the last drawn square before the hole to the center of
the first one after the hole". The problem with that was that the actual
hole size would be smaller than expected: For example, with a thickness of
5 and a hole size of 8, the actual hole would only be 8 - 5 = 3 Kuxels.

This commit aims to fix that by applying a padding to the desired hole
sizes and intervals. The padding is (and pretty much must be) equal to the
thickness of the Kurves.