Add edge padding functionality
Edge padding means that the actual playing field, and thereby the total
size of the game including borders, is decreased in size. Its use is
intended for when the playing field takes up exactly the entire monitor,
which can make it hard to distinguish the top, bottom, and left edges on
some monitors or for example on a projector. With edge padding, the
borders will just be visible inside the boundaries of the monitor.

One notable example of a real life scenario in which edge padding is
useful is when the game is run on a 1440p (= 480*3) monitor with no bezel
at the edge of the actual screen (such as the Asus ROG Swift PG278Q).