Solve Chrome scaling issues hackily
So, as of version 50, Chrome suddenly would not scale Kurve up without
utterly destroying each and every piece of it. It was not even consistent,
but the border would be smoothly upscaled instead of sharply, and the text
would be completely unreadable.

Tried several solutions including `transform: scale3d` and
`backface-visibility: hidden`, and when I got the border right, the canvas
would be blurry instead. I just could not get it right.

This hacky solution relies on the fact that most browsers support the
non-standard `zoom` property. Firefox does not, so we use `-moz-transform:
scale` for it.

`transform: scale` will be the prefered solution if it starts working
again in the future.