Restructure game quit handling
Before this commit, we would reload the entire document when the user
requested to quit the game (and return to the lobby). That works good when
hosting the game locally, but is way too slow when playing online. As of
this commit, the user can quit to the lobby without reloading anything.

When the user quits the game, a new Game is created and the old one is not
used anymore.