Add Z and X keys for Yellow
Ctrl and Alt may not be comfortable for everybody, since they usually have
one or two keys behind them. Furthermore, these keys, especially the
Windows key, may be accidentally pressed by the user, leading to loss of
input focus.

As of this commit, Z and X act as alternative keys for Yellow.