Implement settings functionality and improve GUI
As of this commit, the settings form works: Changes are saved and applied
when closing it and saved between page loads.

Multichoice preferences are now represented as radio buttons, since select
boxes do not work with CSS transform in Firefox.

Settings now have descriptions which are shown when hovering over them in
the settings menu. These are part of the preferences backend.

The "Prevent spawnkills" settings is a dummy and does not work yet.