Add simple unload blocker
Just the standard dialog on beforeunload. Triggering it may sometimes
cause one or more Kurves to not draw anything for ~20px. No obvious
pattern can be seen.

The only reason for this behavior that I can conceive is MainLoop
panicking, but I am not sure how one could solve it. Anyway, I think it is
better than accidentally unloading the entire game due to a misclick.