Add error handling for localStorage access denied
Before this commit, if the browser was set not to allow access to
localStorage (e.g. "Block third-party cookies and site data" in Chrome),
things would crash, the settings menu would not show etc.

This commit adds error handling for this and displays a message telling
the user that settings could not be loaded/saved. Settings can also be
changed and used as expected, they will just not be saved between reloads.