Improve localStorage error handling
PreferenceManager now throws all errors it catches when trying to access
localStorage, not just SecurityError. `get` has been replaced with
`getSaved` to indicate it tries to get from localStorage.

It is recommended that a call to `PreferenceManager.getSaved` is wrapped
in a `try` clause and that `PreferenceManager.getCached` is used in the
`catch` clause as a fallback.