Add spawnkill prevention
It sometimes happens in Kurve that two players spawn close enough to each
other that it is not even theoretically possible for (at least) one of
them to avoid a collision at round start. The spawnkill prevention setting
added by this commit greatly reduces the risk of such situations by
enforcing a minimum distance between players' spawn points.

The minimum distance is normally based on the turning radius and Kurve
thickness, but it is capped to a value such that the circular areas
"reserved" by all Kurves together cannot cover more than half the
available spawn area, so as to avoid sampling for a valid spawn point for
an absurdly long (or infinite) time.

All constants related to the minimum distance can be configured in code.