Dialogs, confirm quit and new event handling
This commit adds the ability to show dialogs, including alerts and
confirms, with the help of `GUIController`. This new functionality made it
necessary to rewrite the event handling, which also became simpler overall
in the process.

As of this commit, the dialog functionality is used to give the player(s)
the optional requirement to confirm their quit action before actually
quitting the game, so as to avoid deleting an entire match by accidentally
hitting the Escape button (which, of course, has _never_ happened to yours
truly ...).