Fix key/mouse handling when settings visible
When the settings menu was being shown, the mouse and keyboard would not
work as expected; for example, neither could be used to modify the value
of an input[type="number"] field. Instead, a mouse click would be
interpreted as Blue getting (un)ready, which is not a desired behavior in
the settings menu.