Add loss of focus handling (Alt + Tab etc)
One problem with running Kurve in the browser is that if the application
loses focus, usually due to the user alt-tabbing, the game cannot detect
if keys/mouse buttons are released, so all Kurves just keep turning like
they were when focus was lost, even if focus is gained again!

As of this commit, if focus is lost, all keys and mouse buttons are
automatically released (from the application's perspective), resulting in
expected and acceptable behavior.