Add allow blurry scaling setting
Since Kurve has a fixed size of 640x480 and only scales natively to
integer multiples of that, players with certain screen resolutions (such
as 1600x900) will get an very small actual game in the middle of a vast
gray background.

This commit adds the option to let the game scale to 1.25 1.5, 1.75 and
2.5 times the original size, at the expense of pixelmapping.

The body class that has been added to <body> is there for specificity
reasons: Without it, `body.blurry` takes precedence over `body`, which in
practice means that integer multiple scales (e.g. 2x) do NOT work if
blurry scaling is allowed. Instead, it goes 1.5x, 1.75x, 2.5x (skipping
2x) and then never changes above that.