Connect setting and description visually
During testing, it was discovered that some users may not see the
description while exploring the settings menu. This commit seeks to make
the description more visible by adding a line connecting the currently
hovered-upon setting with its description, and by making the description
fade in instead of just instantly appearing.