Merge MultichoicePreference values and labels
Prior to this commit, MultichoicePreference would take an array of values
(keys) and an array of labels, and it was the programmer's responsibility to
make sure the nth element of values corresponded to the nth element of labels.
This is for obvious reasons a maintainability nightmare, so as of this commit,
one single array is used instead.