Fix scaling on HiDPI screens
Up until this commit, scaling did not work as intended on HiDPI screens
such as those of the MacBook Pro 13 (2560x1600) and 15 (2880x1800). The
optimum on those resolutions, in theory, is to scale by a factor 3, to
1920x1440.

But that is without considering how those systems actually interpret px
values. The MacBook Pro 13, for example, reports a resolution of
1280x800 to the browser, and that is what our media queries target. So
we get no scaling at all and an incredibly miniscule game.

The solution implemented in this commit is to scale by a factor 1.5 when
the `resolution` (or `device-pixel-ratio`) is 2 and a reported
resolution corresponding to 1.5x scaling is detected (in which case the
physical resolution corresponds to 3x scaling).

For example, on the MacBook Pro 13, the game is scaled 1.5x from 640x480
to 960x720, which covers as much of the screen as 1920x1440 would have
done had the system been running without any HiDPI scaling (i.e.
`device-pixel-ratio` = 1).

We cannot guarantee perfect pixel mapping this way, but it looks good
enough, and it is far better than no scaling at all.