Fix popup/history issues
This is a best-effort attempt at creating a smooth experience for
everyone, including Firefox users, while still trying to avoid unloading
or pausing the game, for example by accidentally clicking the back
button on the mouse.

Before this commit, pressing Space did not take the user to the lobby in
Firefox, but instead displayed a warning which was pretty hard to spot.