Improve round state handling
This commit makes explicit the fact that there is a current round being
(re)played when and only when the `Tick` message is being subscribed to.

The `currentRound` field in `Model` is therefore removed, the `Live` and
`Replay` game states are grouped together under the `MidRoundState`
type, and `Tick` is given a `MidRoundState` parameter.

💡 `git show --color-words='\w+|.'`