Refactor and unify input handling
Instead of keeping track of so-called emulated pressed buttons when
replaying a round, we now compute the effective pressed buttons (which
control how the Kurves actually turn) in the exact same way in both live
and replay mode: namely from the history of user interactions.

This happens to fix a bug where if a player would start holding a button
during the spawn sequence (after pressing Space to start the [next]
round but before the Kurves start moving) and that round would later be
replayed, their turning at the beginning of the round would be ignored
in the replay, because the set of pressed buttons at round start would
have been locked in by `prepareLiveRound` when Space was pressed.

The current solution may not be very optimized, since in every tick we
go through the entire history of user interactions up to the current
tick, but it's good enough for now.

💡 `git show --color-words='==|Recent|(previous|emulated)PressedButtons|case|.'`