Reset players instead of using spawned players
This should make it easier to track user interactions individually for
each player, as each player's interaction history will automatically be
preserved when they are reset.

We sort the players because their order in the finished round depends on
the order in which they died, not in which they spawned.

💡 `git show --color-moved --color-moved-ws=allow-indentation-change`