Track input history per player
Today, each raw button press is stored in the history of the current
round. During replay, the recorded history of button presses is used to
reproduce the players' movements.

As of this commit, each player instead has its own history of new
turning states. This decouples what happened in a recorded round from
the specific keybinds that happened to be used when recording it.

Note that the implementation in this commit records a new turning state
even if doing so is redundant: for example, `TurningRight` may be
recorded for a player even if that is already the newest turning state
in that player's history. This means that every time _any_ player does
_anything_, other players may have redundant entries added to their
history. A condition preventing such redundancy is trivial to add, but
is left out for now because this commit is large enough as it is.

💡 `git show --color-words='computeTurningState .|recordUserInteraction direction button tick|\w+|.'` (not useful for `recordUserInteraction`)