Fix incorrect input handling during spawn
Pressing a button while in the lobby (i.e. before Space is pressed to
initiate spawning), and then holding it, works as expected: the player
starts turning immediately when the players are released.

Pressing a button while the players are spawning (i.e. after Space is
pressed but before the players are released), however, doesn't: the
player starts turning one tick too late.

This is most easily demonstrated by

  * removing all players except Red,
  * setting the turning radius to 250,
  * setting the number of flicker ticks to 20,

and then observing where the Kurve hits the wall when 1 (Red's left key)
is pressed before versus during spawning.

The diff is somewhat unintuitive: in the grand scheme, it is actually
the `PreRound` case of `howToModifyRound` that's fixed by this commit,
but that case itself is locally correct. The real cause of the problem
is that `recordInteractionBefore` is incorrectly implemented (and that
the semantics of `recordUserInteraction` is unintuitive). The `MidRound`
case was locally incorrect in a way such that the off-by-one error in
`recordInteractionBefore` was negated for that case.

💡 `git show --color-words='\w+|.'`