Refactor overlay rendering (#40)
The `clearOverlay` port is only ever used to clear the entire overlay
canvas (i.e. never a specific region). This PR moves that functionality
into the `renderOverlay` port, making it more akin to Elm's `view`
function (i.e. it is now used to describe the entire desired state of
the overlay canvas, rather than any additions to it).

As of this PR, clearing the overlay is simply done by rendering the
empty list of squares. This is necessary when starting a new round,
because otherwise the winning player's head is incorrectly shown while
the players are spawning, but it's not necessary in `update` anymore,
because it's implicitly done by the port itself.

This PR was made possible by #39.

💡 `git show --color-words='\w+|.'`

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>