Add real lobby (#57)
This PR replaces the placeholder lobby added in #44 with an actual lobby
where players have to actively join (by pressing their left key) in
order to participate in the game. As in the legacy JavaScript version,
the game can only be started if at least one player is ready. (This is a
deliberate difference from the original MS-DOS version, which requires
at least two players; the purpose is to allow players to practice on
their own.)

The PNG images are derived from screenshots of the original MS-DOS
version, via the legacy JavaScript version. For this PR, they have been
turned completely black (since the color is now specified using CSS) and
optimized with `optipng -strip all -o7`.

The various elements of the lobby are placed so as to pixel-perfectly
match the legacy JavaScript version; hence the seemingly arbitrary `px`
dimensions in the stylesheet.

💡 `git show --color-words='\[ Html\.text.+|\w+|.'`

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>