Add scoreboard (#58)
As in the original game, all living players are awarded a point whenever
someone else dies. If a round is replayed, so is the scoreboard.

The score-based game-over condition from the original game will be added
later, together with the traditional "KONEC HRY" end screen with the
final results. The digits added in this PR are therefore implemented
with support for small digits in mind.

For simplicity, `combineScores` is written with the assumption that each
participating player will always be represented in both of its inputs,
even though that isn't captured by the types.

💡 `git show --color-words='\{ model.+|\w+|.'`

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>