Simplify scaling/zoom philosophy (#67)
The goal of the current solution was to automatically scale the game up
as much as possible while ensuring perfect pixel mapping. While it
worked fairly well in many scenarios, there are several reasons to
abandon it:

* It's extremely complicated. Not only is it inherently difficult to
extract all the necessary information and express the intended behavior
– the implementation itself is unacceptably impractical. The idea was to
use `dev/css-generator.js` (introduced before we had SASS) to generate
CSS code and manually paste it into the stylesheet. That required
several manual and implicit steps (including calling the `generateCSS`
function, printing its return value and replacing the correct part of
the stylesheet with it), and of course the stylesheet eventually went
out of sync with the generator (in
`a6e7671582f0fd8cb25e7d51cd5fedcc67788df2`).

* It results in unintuitive behavior when zooming: the game grows
gradually until it doesn't fit anymore, at which point it suddenly
shrinks to half the height and width. This both prevents the user from
zooming in as much as they might want and makes it hard for them to
understand what approximate zoom level they're currently at.

* It doesn't work as intended in Firefox anymore: the layout breaks when
the size of the viewport doesn't allow centering the canvas without the
scoreboard ending up partially off-screen (e.g. 1400px × 1000px).

* Pixel mapping can't be guaranteed anyway: users may have their OS set
to 125% scaling and their browser zoomed to 190%, or whatever.
`image-rendering: pixelated` is good enough for users who don't care,
and those who do can ensure pixel mapping themselves.

The new solution doesn't attempt to ensure perfect pixel mapping or
automatically scale anything up. Instead, we simply declare that we want
images to be pixelated and expect the user to zoom in as they see fit.
The canvas is still centered as long as scoreboard visibility isn't
compromised. The pathological case of a viewport smaller than
640px × 480px (the native size of the original game) isn't supported in
any particular way.

The ID and class on the body are removed because this PR makes them
obsolete. (If I recall correctly, they were there for specificity
reasons.)

💡 `git show --color-words=.`