Only render canvases when in-game (#72)
This PR decouples the overall view of the application from the canvases,
which are only really needed when in-game (not e.g. in the lobby).

Without the changes in `index.html`, the application immediately crashes
when the first round is started:

Uncaught TypeError: Cannot read properties of null (reading
'getContext')

This happens because we clear the canvases at the start of every round,
including the first round, when, as of this PR, they don't yet exist.
"Why isn't `drawSquare` modified then?" Because it's only used when the
canvases do exist, and should we ever try to draw on a canvas that
doesn't exist, we probably don't want that to fail silently. _Clearing_
a canvas that doesn't exist, however, probably isn't that big of an
issue, because it's probably about to be created/replaced, so it'll end
up empty anyway.