Make menu screens borderless (#74)
This PR removes the gray border on the so-called menu screens (i.e. the
lobby, the end screen and the not-yet-existing splash screen), making
them all black, as in the original MS-DOS game. The scoreboard is
refactored accordingly, reflecting the fact that it's now only visible
when in-game.

Notably, this PR makes our clone virtually indistinguishable from the
original MS-DOS game when run in a 640px × 480px viewport (except that
the splash screen doesn't exist yet).

💡 `git show --ignore-space-change --color-words='List \(Html msg\)|\w+|.'`