Lobby font identified
It's most likely the default font of Borland Graphics Interface.

https://en.wikipedia.org/wiki/Borland_Graphics_Interface

Docs:

The default 8x8 bitmapped font is built into the graphics package, so it's
always available at run time. The stroked fonts are each kept in a separate
.CRR file; they can be loaded at run time or converted to .OBJ files (with
the BGIOBJ utility) and linked into your .EXE file.

http://bitsavers.informatik.uni-stuttgart.de/pdf/borland/borland_C++/Borland_C++_Version_4.0_DOS_Reference_Oct93.pdf

So the default font is built in and the splash screen/scoreboard/KONEC
HRY font is stored in TSCR.CHR