Add original game for reference (#80)
It's important to have the original game close at hand, as it's often
needed for reference.

In addition to running the game as it is, one can easily modify
`ZATACKA.EXE` to show any (short) ASCII string using either of the two
fonts used in the game. Example:

    sed -i "s/Achtung, die Kurve/Let's play Achtung/" docs/original-game/ZATACKA.EXE

Note that the replacement string must have the same length (in bytes) as
the original; otherwise, the game crashes.