Add general GUI text support (#78)
This PR replaces all static images containing text using the Borland
Graphics Interface default bitmap font – i.e. the player controls,
"READY", and the digits used to show the results at the end of a game –
with a general text engine that can render any ASCII string.

All glyphs are stored in a single monochrome transparent PNG file, where
each character is offset by 8 multiplied by its character code. (Note
that not every glyph "starts" at the left edge of its 8 × 8 bounding
box.) That file is then used as a clipping mask together with a
background color, just like how we already render large digits since
#58.

How to extract glyphs from the original game:

  1. Run `node tools/extract-glyphs-from-original-game.js docs/original-game/ZATACKA.EXE "(L.Arrow D.Arrow)"`.
  2. Each time the game is launched, go to the lobby and take a
     screenshot (Ctrl + F5), then close DOSBox.
  3. Use your image editing software of choice to create a "sprite map"
     based on the screenshots.
  4. Optimize the image with `optipng -strip all -o7`.

Notably, the mapping from player IDs to controls is exactly as hardcoded
and partial (in the sense that an ID not in [0, 5] results in no text
being shown) as it was before this PR.

💡 `git show --ignore-space-change --color-words='\( left, right \)|controls |Attr\.\w+|digits-small|(line-height|vertical-align): .+$|\w+|.'`