Pause game when window focus is lost (#86)
It's fairly easy to accidentally switch tabs or similar in the middle of
a round. The legacy JavaScript version automatically pauses the game
when that happens, but it doesn't indicate in any way that the game is
paused or how to continue, so it feels like it has frozen.

This PR ports the feature to Elm, this time with the text "Press Space
to continue" being shown when the game is paused. The text is
semi-transparent because players must be able to see in what direction
they are moving and what's going on in their immediate vicinity, in
order to be as prepared as possible when the game is resumed.

💡 `git show --color-words=.`