Fix broken GUI text in Webkit browsers (#89)
Turns out the "READY" text in the lobby is utterly broken in Chrome: it
shows up as `$%      23`, and the characters are 8 × 8 instead of
16 × 16 and repeated in a visually weird fashion. The same goes for the
"Press Space to continue" text and the end-screen scoreboard.

The root cause is missing `mask-size` support – adding a `-webkit-`
fallback makes everything render correctly. According to
https://caniuse.com, this should be enough to support all major browsers
at least back to 2017. (I'm also adding a fallback for `mask-repeat`, of
course, but that doesn't have any visible effect as long as we only try
to render supported characters.)