Fix canvas clearing issue (#90)
In Firefox on Ubuntu on my ThinkPad X1 Carbon 5th Gen, holes are
consistently "invisible" (there appears to be no hole, even though there
actually is) and Kurves are consistently not cleared when proceeding to
the next round. After a confusing debugging session, Kryddan and I
concluded that when the `clearRect` method is called with very large
values for the width and height, the canvas isn’t cleared. I have
reported it as a bug in Firefox:

https://bugzilla.mozilla.org/show_bug.cgi?id=1828137

I have only encountered the issue in Firefox 109–112 (and all the way
back to 106 with `mozregression`; see Bugzilla thread) on said ThinkPad
with a Core i7-7500U running Ubuntu 20.04.1 with Linux 5.15.0. It
doesn't happen in Chrome on Ubuntu, Firefox on Windows 10 (on the same
laptop), or Firefox on Ubuntu 20.04.4 with Linux 5.4.0 on my main PC
with an RTX 2060 Super.

The reason I haven't noticed such an obvious bug until now is that it
didn't manifest itself until I did `sudo apt update && sudo apt upgrade`
a week or two ago and (apparently) got a new version of Firefox.

💡 `git show --color-words='[A-Za-z._]+|.'`