Add confirmation dialog when quitting (#83)
In the original MS-DOS game, when Escape is pressed between rounds, the
game immediately exits to the splash screen. In
83ca2595037f462f6467d196d15ae146aaeec5b9, the legacy JavaScript version
introduced an optional confirmation dialog (enabled by default), because
accidentally quitting in the middle of a game _sucks_. Unlike in the
original game, "quitting" in the legacy JavaScript version means going
to the lobby, not the splash screen, because that's almost always more
useful.

This PR brings a similar confirmation dialog to the Elm version. Since
we don't have any user-facing preferences (yet), it's always enabled.

Having experimented with various approaches, I've decided to go with
this one, which implements keyboard interaction entirely in Elm, instead
of relying on the browser's native behavior.

Neither solution is without its flaws. The browser approach has these
disadvantages:

  * The arrow keys don't work for moving focus between the buttons.
  * Pressing Tab (or Shift + Tab) cycles through _all focusable
    elements_ in the page, and even the browser chrome, so the dialog
    doesn't feel modal.
  * When the dialog is opened, the "No" button must be focused with
    `Browser.Dom.focus` and `Task.attempt`. That is, we must look it up
    by ID and handle the possibility that it doesn't exist, even though
    we know that it does.
  * We have to control in Elm whether we should call the
    `preventDefault` method in JavaScript or not (for example by
    conditionally setting an attribute in Elm and check for it in
    JavaScript).

The Elm approach has these:

  * We're reinventing the wheel when it comes to moving focus between
    input elements and interacting with them.
  * Supporting more than two input elements would require a more general
    implementation for the complexity to stay reasonable.
  * Since we use `event.code` (the physical key), not `event.key` (the
    meaning of the key), the dialog may not behave as expected with
    non-QWERTY layouts.

While we suspect that the Elm approach may not cut it if/when we
introduce more input elements down the road, I think this implementation
is good enough for now. I also believe there's value in documenting in
the Git history that we initially did it this way and why.

I decided to go with `<div>` rather than `<dialog>` because [the latter
isn't supported in relatively new versions of Firefox (97) and Safari
(15.3)](https://caniuse.com/dialog).

💡 `git show --color-words='goToLobby .+|\w+|.' --ignore-space-change`