Upgrade elm-review to v2.14.0 (#98)
I did this:

    npm install elm-review@latest --save-exact --save-dev

Then I got version 2.12.0 of the `elm-review` CLI, with which `npm run review` failed with this error:

    -- UNSUPPORTED ELM-REVIEW VERSION ----------------------------------------------
    
    You are using an unsupported version of the jfmengels/elm-review Elm package.
    You are using version 2.10.0, but I need it to be 2.14.0 <= v < 3.0.0.
    
    Please upgrade your version by running the following commands:
    
    cd /…/kurve/review
    npx elm-json install jfmengels/elm-review@2
    
    If that doesn't work, try out:
    
    cd /…/kurve/review
    npx elm-json upgrade --unsafe

I ran the first suggested command, but it only upgraded to 2.13.1, so `npm run review` still failed with the same error (modulo the version it said I was using, of course).

So I tried the "unsafe" command instead, which made `npm run review` work. But it then reported this error:

    (fix) Simplify: List.any with a check for equality with a specific value can be replaced by List.member with that value

I fixed it with `npx elm-review --fix`.