Simplify writing tests for how rounds should end (#107)
I think we're going to want to write quite a few test cases about how a round should end given some state. This PR adds an `expectRoundOutcome` function that takes care of advancing the round until the expected final tick and then expect it to end in the specified way.

The new `Spawning _ _` case (which is implicitly covered by the deleted `_` case) is a bit weird: a round should never re-enter that state, so why do we even have to consider it as a possible outcome of a tick? Maybe something to improve going forward.

💡 `git show --ignore-space-change`