Improve test case for crashing into wall (#108)
Today, the test case doesn't verify that death occurs exactly at the time of collision. It would pass even if the implementation had a bug that killed the Kurve too early (so that it wouldn't be alive at the position it has at the start of the test).

This PR makes the test case verify that the Kurve dies exactly when it crashes into the wall.

I chose 2.5 instead of 2 as the x coordinate because then the pixel-aligned head of the Kurve is perfectly centered (along the x axis) around its "actual" position. I think this may become particularly relevant if we get around to redesigning collision handling to not be pixel-based.

💡 `git show --color-words='0.5|.'`