Use properly initialized rounds in tests (#115)
Manually constructing `Round`s like we do in test cases can result in really confusing situations. For example, consider this test case (without the changes in this PR):

```elm
myTest : Test
myTest =
    test
        "A Kurve that crashes into another Kurve's tail dies"
        (\_ ->
            let
                round : Round
                round =
                    { kurves =
                        { alive = [ red, green ]
                        , dead = []
                        }
                    , occupiedPixels = Set.empty
                    , initialState =
                        { seedAfterSpawn = Random.initialSeed 0
                        , spawnedKurves = [ red, green ]
                        }
                    , seed = Random.initialSeed 0
                    }
            in
            round
                |> expectRoundOutcome
                    { tickThatShouldEndIt = tickNumber 8
                    , howItShouldEnd = always Expect.pass
                    }
        )


red : Kurve
red =
    let
        kurveState =
            { position = ( 100, 100 )
            , direction = Angle 0
            , holeStatus = Unholy 60
            }
    in
    { color = Color.red
    , id = 0
    , controls = ( Set.empty, Set.empty )
    , state = kurveState
    , stateAtSpawn = kurveState
    , reversedInteractions = []
    }


green : Kurve
green =
    let
        kurveState =
            { position = ( 98, 110 )
            , direction = Angle (pi / 2)
            , holeStatus = Unholy 60
            }
    in
    { color = Color.green
    , id = 3
    , controls = ( Set.empty, Set.empty )
    , state = kurveState
    , stateAtSpawn = kurveState
    , reversedInteractions = []
    }
```

It seems to describe a round in which Red spawns at (100, 100) and Green spawns at (98, 110). One would expect Green to crash into Red's tail end, so the test should pass. But it fails.

Turns out the pixels that one would expect to be occupied by Red's tail end are _not_ occupied, so Green just keeps going. This happens because we set `occupiedPixels` to `Set.empty`, i.e. it doesn't include the pixels representing Red's tail end.

In a real game, `occupiedPixels` is always initialized based on the Kurves' spawn positions; see `prepareRoundHelper`. This PR makes the test suite use that function as well, and renames it to the not-so-internal name `prepareRoundFromKnownInitialState`.

💡 `git show --color-words='[A-Z]?[a-z]+|.'`